requirements.txt
alembic
mysqlclient
python-dotenv
SQLAlchemy
